Cleanup, move types, add comments
